Adapter: Replace protocol interface by protocol enum again
