+++ adapter/src/main/java/org/jvalue/ods/adapterservice/adapter/Protocol.java
@@ -1,7 +1,18 @@
 package org.jvalue.ods.adapterservice.adapter;
 
+import org.jvalue.ods.adapterservice.adapter.importer.HttpImporter;
 import org.jvalue.ods.adapterservice.adapter.importer.Importer;
+import org.springframework.web.client.RestTemplate;
 
-public interface Protocol {
-    Importer getImporter();
+public enum Protocol {
+  HTTP(new HttpImporter(new RestTemplate()));
+
+  private final Importer importer;
+  Protocol(Importer importer) {
+    this.importer = importer;
+  }
+
+  public Importer getImporter() {
+    return importer;
+  }
 }
